Changed the way Board and Point is represented. Empty cell (no color) is now "0" instead of "-1". Point represented by unsigned ints instead of ints.

+++ test/TestTools.cpp
@@ -4,7 +4,7 @@
 
 namespace TestTools
 {
-Board createBoard(const std::vector<std::vector<int>>& data)
+Board createBoard(const std::vector<std::vector<unsigned int>>& data)
 {
     const unsigned int columnCount{static_cast<unsigned int>(data.size())};
     const unsigned int rowCount{static_cast<unsigned int>(data[0].size())};
@@ -13,8 +13,7 @@ Board createBoard(const std::vector<std::vector<int>>& data)
     Board board(columnCount, rowCount, stringStream);
     for (unsigned int row = 0; row < rowCount; ++row)
         for (unsigned int column = 0; column < columnCount; ++column)
-            board.setColor({static_cast<int>(column), static_cast<int>(row)},
-                           data[column][row]);
+            board.setColor({column, row}, data[column][row]);
     return board;
 }
